refactor(component,module):  servicesフォルダ統一 #133
