+++ src/app/profile/profile/profile.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, HostListener } from '@angular/core';
 import { FormBuilder, Validators, FormControl } from '@angular/forms';
-import { UserProfileService } from 'src/app/service/user-profile.service';
+import { UserProfileService } from 'src/app/services/user-profile.service';
 import { AuthService } from 'src/app/services/auth.service';
-import { JobPostService } from 'src/app/service/job-post.service';
+import { JobPostService } from 'src/app/services/job-post.service';
 @Component({
   selector: 'app-profile',
   templateUrl: './profile.component.html',
