+++ src/app/services/company-profile.service.ts
@@ -4,7 +4,7 @@ import { CompanyProfile } from '../interfaces/profile';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
